requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0

# Database
pymongo==4.6.1

# Machine Learning & Data Processing
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.0
geopandas==0.14.1
shapely==2.0.2

# Visualization
matplotlib==3.8.0
seaborn==0.13.0

# Google Cloud & AI
google-cloud-aiplatform==1.36.4
vertexai==0.0.1
google-cloud-storage==2.13.0

# Bayesian Network
pgmpy==0.1.24

# Development & Testing
pytest==7.4.2
black==23.9.1
flake8==6.1.0

# Additional utilities
tqdm==4.66.1
requests==2.31.0
python-dateutil==2.8.2 